Add files via upload
This is a project I made when I was a data analyst at Salon Paris. 

It was my job to go through the lists of data that had been collected in the past couple of weeks. I calculated some important metrics, that Salon Paris can use to plan out the operation of the business for the rest of the month.